+++ pg-client/src/pg/client.clj
@@ -4,17 +4,14 @@
    [clojure.string :as str]
    [pg.client.conn :as conn]
    [pg.client.copy :as copy]
-   [pg.client.func :as func]
    [pg.client.quote :as quote]
-   [pg.client.result :as res]
+   [pg.client.flow :as flow]
    [pg.client.sql :as sql]
    [pg.const :as const]
    [pg.oid :as oid]
    [pg.types.hint :as hint])
   (:import
    clojure.lang.Keyword
-   clojure.lang.PersistentVector
-   java.util.ArrayList
    java.util.HashMap
    java.util.List
    java.util.Map
@@ -100,7 +97,7 @@
 
      (conn/describe-statement conn statement)
      (conn/send-sync conn)
-     (res/interact conn :prepare init))))
+     (flow/interact conn :prepare init))))
 
 
 (defn close-statement
@@ -112,7 +109,7 @@
         stmt]
     (conn/close-statement conn statement))
   (conn/send-sync conn)
-  (res/interact conn :close-statement))
+  (flow/interact conn :close-statement))
 
 
 (defmacro with-statement
@@ -137,7 +134,7 @@
   "
   [conn]
   (conn/authenticate conn)
-  (res/interact conn :auth)
+  (flow/interact conn :auth)
   conn)
 
 
@@ -259,7 +256,7 @@
 
   ([conn sql opt]
    (conn/send-query conn sql)
-   (res/interact conn :query opt)))
+   (flow/interact conn :query opt)))
 
 
 (defn execute-statement
@@ -306,7 +303,7 @@
      (conn/send-execute conn portal rows)
      (conn/close-portal conn portal)
      (conn/send-sync conn)
-     (res/interact conn :execute opt))))
+     (flow/interact conn :execute opt))))
 
 
 (defn execute
@@ -544,113 +541,3 @@
           null const/COPY_CSV_NULL
           format const/COPY_FORMAT_CSV}}]
    (copy/copy-in-maps conn sql maps keys format oids sep end null)))
-
-
-;;
-;; Reducers
-;;
-
-(def first
-  "
-  Return the first value skipping all the rest rows.
-  "
-  {:fn-init (fn [] [])
-   :fn-reduce (fn [^PersistentVector acc x]
-                (if (.isEmpty acc)
-                  (conj acc x)
-                  acc))
-   :fn-finalize clojure.core/first})
-
-
-(def java
-  "
-  Return an ArrayList of HashMaps.
-  "
-  {:fn-keyval func/zipmap-java
-   :fn-init #(new ArrayList)
-   :fn-reduce (fn [^ArrayList array-list row]
-                (doto array-list (.add row)))})
-
-
-(def kebab-keys
-  "
-  Return a vector of Clojure maps where the keys
-  transformed to `:kebab-case-keywords`.
-  "
-  {:fn-column func/kebab-keyword})
-
-
-(def matrix
-  "
-  Return a vector of vectors of values.
-  "
-  {:fn-unify func/unify-none
-   :fn-keyval func/vals-only
-   :fn-init #(transient [])
-   :fn-reduce conj!
-   :fn-finalize persistent!})
-
-
-(defn index-by
-  "
-  Return a Clojure map of {(f row) => row}. The function
-  can be a keyword, for example for `:id`, you'll get a map
-  like `{1 {:id 1, name '...'}, 2 {:id 2, name '...'}}`.
-  "
-  [f]
-  {:fn-init #(transient {})
-   :fn-reduce (fn [acc! row]
-                (assoc! acc! (f row) row))
-   :fn-finalize persistent!})
-
-
-(defn group-by
-  "
-  Return a map of {(f row) => [row1, row2, ...]}. The function
-  can be a keyword, for example for `:name`, you'll get a map
-  like `{'ivan' [{:id 1, ...} {:id 2, ...}]}`.
-  "
-  [f]
-  {:fn-init hash-map
-   :fn-reduce (let [-conj (fnil conj [])]
-                (fn [acc row]
-                  (update acc (f row) -conj row)))})
-
-
-(defn kv
-  "
-  Return a map of {(fk row) => (fv row)}.
-  The `fk` is a key function, and `fv` is a value function.
-  Both functions accept the same row.
-  "
-  [fk fv]
-  {:fn-init #(transient {})
-   :fn-reduce (fn [acc! row]
-                (assoc! acc! (fk row) (fv row)))
-   :fn-finalize persistent!})
-
-
-(defn run
-  "
-  Run the `f` function for each row in the result.
-  Return a number of rows processed. For side effects
-  only.
-  "
-  [f]
-  {:fn-init (constantly 0)
-   :fn-reduce (fn [n row]
-                (f row)
-                (inc n))
-   :fn-finalize identity})
-
-
-(defn fold
-  "
-  Reduce the rows using the init value (an accumulator)
-  and a function that accepts the current accumulator and
-  the next row. The function should return the new accumulator.
-  "
-  [init f]
-  {:fn-init (constantly init)
-   :fn-reduce f
-   :fn-finalize identity})