+++ pg-client/test/pg/client_test.clj
@@ -18,6 +18,7 @@
    [clojure.string :as str]
    [clojure.test :refer [deftest is use-fixtures testing]]
    [pg.client :as pg]
+   [pg.client.as :as as]
    [pg.client.conn :as conn]
    [pg.client.func :as func]
    [pg.honey :as pgh]
@@ -1153,7 +1154,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as pg/java})]
+          (pg/execute conn query nil {:as as/java})]
 
       (is (= [{:b 2 :a 1}
               {:b 4 :a 3}
@@ -1174,7 +1175,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as (pg/index-by :a)})]
+          (pg/execute conn query nil {:as (as/index-by :a)})]
 
       (is (= {1 {:a 1 :b 2}
               3 {:a 3 :b 4}
@@ -1191,7 +1192,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as (pg/group-by :a)})]
+          (pg/execute conn query nil {:as (as/group-by :a)})]
 
       (is (= {1 [{:a 1 :b 2}]
               3 [{:a 3 :b 4}]
@@ -1207,7 +1208,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as (pg/kv :b :a)})]
+          (pg/execute conn query nil {:as (as/kv :b :a)})]
 
       (is (= {2 1
               4 3
@@ -1226,7 +1227,7 @@ drop table %1$s;
           (atom [])
 
           as
-          (pg/run
+          (as/run
             (fn [row]
               (swap! capture! conj row)))
 
@@ -1247,7 +1248,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           as
-          (pg/fold #{} (fn [acc {:keys [a b]}]
+          (as/fold #{} (fn [acc {:keys [a b]}]
                          (conj acc [a b])))
 
           res
@@ -1264,7 +1265,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as pg/matrix})]
+          (pg/execute conn query nil {:as as/matrix})]
 
       (is (= [[1 2]
               [3 4]
@@ -1280,7 +1281,7 @@ drop table %1$s;
           "with foo (a, b) as (values (1, 2), (3, 4), (5, 6)) select * from foo"
 
           res
-          (pg/execute conn query nil {:as pg/first})]
+          (pg/execute conn query nil {:as as/first})]
 
       (is (= {:a 1 :b 2} res)))))
 
@@ -2059,7 +2060,7 @@ copy (select s.x as X from generate_series(1, 3) as s(x)) TO STDOUT WITH (FORMAT
           (pgh/query conn
                      {:select [[[:inline "string"] :foo]]}
                      {:pretty true
-                      :as pg/first})]
+                      :as as/first})]
       (is (= {:foo "string"} res)))))